refactor(#30): move /domain/entities to /models
